requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
chromadb
PyMuPDF